audio-analysis-function/requirements.txt
functions-framework==3.*
google-cloud-storage==2.10.0
google-cloud-firestore==2.11.1
librosa==0.10.1
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1